Never mind the lookup
That was a bad place for a lookup.  It caused the
dictionary to blow up by a factor of 5 and slowed things
down by a factor of two (for 6,5).  The moral of the story
is that you only want to capture choice nodes that are
part of the tree.

Well and good, but I don't see how my code here can
avoid scanning the whole tree.  Sure it's compact when
standardized, but we have to visit every branch to get
there.

Well, not branches that standardize to 0, but any branch
that goes to a node gets worked out completely.  If you look
at choice and substitute, they're going to descend all the
way into the tree, with all the duplication that implies.